Extracted RelatedFilesFinder
